+++ tests/routesTests/GetStockTests.py
